whatsdue: moved utility function to util folder
