+++ uqcsbot/util/uq_course_util.py
